proof-producer/benchmarks/requirements.txt
BenchExec==3.27
certifi==2025.1.31
charset-normalizer==3.4.1
Deprecated==1.2.18
googleapis-common-protos==1.68.0
grpcio==1.70.0
idna==3.10
importlib_metadata==8.5.0
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
protobuf==5.29.3
PyYAML==6.0.2
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
zipp==3.21.0
